% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppr.R
\docType{data}
\name{Aeut}
\alias{Aeut}
\title{Oomycete root rot pathogen *Aphanomyces euteiches* AFLP data}
\format{
a [genind()] object with two populations containing a data
  frame in the `other` slot called `population_hierarchy`. This
  data frame gives indices of the populations and subpopulations for the data
  set.
}
\usage{
data(Aeut)
}
\description{
The Aeut dataset consists of 187 isolates of the Oomycete root 
  rot pathogen, *Aphanomyces euteiches* collected from two different 
  fields in NW Oregon and W Washington, USA.
}
\references{
Grunwald, NJ and Hoheisel, G.A. 2006. Hierarchical Analysis of
  Diversity, Selfing, and Genetic Differentiation in Populations of the 
  Oomycete *Aphanomyces euteiches*. Phytopathology 96:1134-1141
  doi: \href{https://doi.org/10.1094/PHYTO-96-1134}{10.1094/PHYTO-96-1134}
}
