% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bitwise.pgen}
\alias{bitwise.pgen}
\title{Calculates and returns a matrix of Pgen values for the given genlight object.
Each column represents a genotype in the genlight objects, and each row
represents a specific, sequential set of 8 base pairs. The values in each cell
represent the Pgen value of the corresponding set of base pairs. These values
indicate the probability of observing these alleles in a randomly mating
population using estimates derived from the genotypes present in the genlight
object.}
\usage{
bitwise.pgen(x, log = TRUE)
}
\arguments{
\item{x}{a genlight object.}

\item{log}{a \code{logical} to determine whether the values should be returned
 as percentages or logarithms of percentages. \code{TRUE} is the default, and
 returns the logarithmic values rather than the percentage values. This option has
 a much larger range and is highly recommended. \code{FALSE} returns the percentage
 chance for each genotype to be produced via random mating, rather than the log
 equivalent.}
}
\value{
A vector containing one Pgen value for each genotype in the genlight object.
}
\description{
Calculates and returns a matrix of Pgen values for the given genlight object.
Each column represents a genotype in the genlight objects, and each row
represents a specific, sequential set of 8 base pairs. The values in each cell
represent the Pgen value of the corresponding set of base pairs. These values
indicate the probability of observing these alleles in a randomly mating
population using estimates derived from the genotypes present in the genlight
object.
}
\author{
Zhian N. Kamvar, Jonah Brooks
}

