% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bitwise}
\alias{bitwise}
\alias{bitwise.dist}
\alias{bitwise.pgen}
\title{Calculate a distance matrix comparing samples based on the number of alleles
that differ in zygosity.}
\usage{
bitwise.dist(x, percent = TRUE, mat = FALSE, missing_match = TRUE,
  threads = 0)

bitwise.pgen(x, percent)
}
\arguments{
\item{x}{a genlight object.}

\item{percent}{\code{logical}. Should the distance be represented from 0 to 1?
Default set to \code{TRUE}. \code{FALSE} will return the distance represented
as integers from 1 to n where n is the number of loci.}

\item{mat}{\code{logical}. Return a matrix object. Default set to
\code{FALSE}, returning a dist object. \code{TRUE} returns a matrix object.}

\item{missing_match}{\code{logical}. Determines whether two samples differing
by missing data in a location should be counted as matching at that location.
Default set to \code{TRUE}, which forces missing data to match with anything.
\code{FALSE} forces missing data to not match with any other information.}

\item{threads}{The maximum number of parallel threads to be used within this
  function. A value of 0 (default) will attempt to use as many threads as there
  are available cores/CPUs. In most cases this is ideal. A value of 1 will force
  the function to run serially, which may increase stability on some systems.
  Other values may be specified, but should be used with caution.}

\item{x}{a genlight object.}

\item{percent}{a \code{logical} to determine whether the values should be returned
 as percentages or logarithms of percentages. \code{FALSE} is the default, and
 returns the logarithmic values rather than the percentage values. This option has
 a much larger range and is highly recommended. \code{TRUE} returns the percentage
 chance for each genotype to be produced via random mating, rather than the log
 equivalent.}
}
\value{
Pairwise distances between individuals present in the genlight object.

A vector containing one Pgen value for each genotype in the genlight object.
}
\description{
This function does pairwise comparisons between diploid samples in a genlight
object. The number representing the distance between two samples is equal to
the number of alleles in the samples that do not have the same zygosity.

Calculates and returns a vector of Pgen values for the given genlight object.
Each element represents the probability that the individual at that element
would have been produced via random mating using estimates derived from the
genotypes present in the genlight object.
}
\author{
Zhian N. Kamvar, Jonah Brooks

Zhian N. Kamvar, Jonah Brooks
}

