% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bitwise.dist}
\alias{bitwise.dist}
\title{Calculate a distance matrix comparing samples based on the number of alleles
that differ in zygosity.}
\usage{
bitwise.dist(x)
}
\arguments{
\item{x}{a genlight object.}
}
\value{
Pairwise distances between individuals present in the genlight object.
}
\description{
This function does pairwise comparisons between diploid samples in a genlight
object. The number representing the distance between two samples is equal to
the number of alleles in the samples that do not have the same zygosity.
}
\author{
Zhian N. Kamvar, Jonah Brooks
}

