% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstraping.R
\name{boot_ci}
\alias{boot_ci}
\title{Perform bootstrap statistics, calculate and plot confidence intervals.}
\usage{
boot_ci(tab, n = 1000, ci = 95, total = TRUE, rarefy = FALSE,
  mlg.weight = TRUE, ...)
}
\arguments{
\item{tab}{a genind object OR a matrix produced from
\code{\link[poppr]{mlg.table}}.}

\item{n}{an integer defining the number of bootstrap replicates (defaults to
1000).}

\item{ci}{the percent for confidence interval.}

\item{total}{argument to be passed on to \code{\link[poppr]{mlg.table}} if
\code{tab} is a genind object.}

\item{rarefy}{if \code{TRUE}, bootstrapping will be performed on the smallest
population size. Defaults to \code{FALSE}, indicating that bootstrapping
will be performed respective to each population size.}

\item{mlg.weight}{when \code{FALSE} (default), all observed MLGs have an equal
chance of being selected in the bootstrapping procedure. When \code{TRUE},
larger MLGs have more of a chance of being selected. This parameter is
ignored when \code{rarefy = TRUE}.}

\item{...}{parameters to be passed on to \code{\link[boot]{boot}} and
  \code{\link{get_stats}}}
}
\value{
an array of 3 dimensions giving the lower and upper bound, the index
  measured, and the population. This also prints barplots for each population
  faceted by each index using \pkg{ggplot2}. This plot can be retrieved by
  using \code{p <- last_plot()}
}
\description{
Perform bootstrap statistics, calculate and plot confidence intervals.
}
\note{
While it is possible to use custom functions with this, there are three
  important things to remember when using these functions:
\enumerate{
\item The function must return a single value.
\item The function must allow for both matrix and vector inputs
\item The function name cannot match or partially match any arguments from
\code{\link[boot]{boot}}
}. Anonymous functions are okay (e.g. \code{function(x) vegan::rarefy(x, 10)}).
}
\examples{
library(poppr)
data(Pinf)
boot_ci(Pinf, n = 100)
\dontrun{
# This can be done in a parallel fasion (OSX uses "multicore", Windows uses "snow")
system.time(boot_ci(Pinf, 10000L, parallel = "multicore", ncpus = 4L))
system.time(boot_ci(Pinf, 10000L))

# The previous version of poppr contained a statistic known as Hexp, which
# was caluclated as (n/(n - 1))*lambda. It basically looks like an unbiased
# Simpson's index. This statistic was originally included in poppr because it
# was originally included in the program multilocus. Since the reference for
# this was hard to track down. Because of this, it was removed from analysis.

Hexp <- function(x){
 lambda <- vegan::diversity(x, "simpson")
 x <- drop(as.matrix(x))
 if (length(dim(x)) > 1){
   N <- rowSums(x)
 } else {
   N <- sum(x)
 }
 return((N/(N-1))*lambda)
}
boot_ci(Pinf, 1000L, Hexp = Hexp)
boot_ci(Pinf, 1000L, Hexp = Hexp, rarefy = TRUE)
}
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{do_boot}} \code{\link{get_stats}} \code{\link{poppr}}
}

