\name{[}
\alias{[}
\alias{dim}
\alias{initialize}
\title{Methods used for the bootgen object.}
\arguments{
  \item{x}{a \code{"\linkS4class{bootgen}"} object}

  \item{i}{vector of numerics indicating number of
  individuals desired}

  \item{j}{a vector of numerics corresponding to the loci
  desired.}

  \item{...}{unused.}

  \item{drop}{set to \code{FALSE}}

  \item{.Object}{a character, "bootgen"}

  \item{gen}{\code{"\linkS4class{genind}"} object}

  \item{replen}{a vector of numbers indicating the repeat
  length for each microsatellite locus.}
}
\description{
  This is not designed for user interaction.
}

