\name{initialize}
\alias{[}
\alias{dim}
\alias{initialize}
\title{Methods used for the bruvomat object.}
\arguments{
  \item{.Object}{a character, "bruvomat"}

  \item{gen}{\code{"\linkS4class{genind}"} object}

  \item{replen}{a vector of numbers indicating the repeat
  length for each microsatellite locus.}

  \item{x}{a \code{"\linkS4class{bruvomat}"} object}

  \item{i}{vector of numerics indicating number of
  individuals desired}

  \item{j}{a vector of numerics corresponding to the loci
  desired.}

  \item{...}{unused.}

  \item{drop}{set to \code{FALSE}}
}
\description{
  This is not designed for user interaction.
}

