% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitwise.r
\name{check_openMP_support}
\alias{check_openMP_support}
\title{Determines whether openMP is support on this system.}
\usage{
check_openMP_support()
}
\value{
FALSE if openMP is not supported, TRUE if it is
}
\description{
Determines whether openMP is support on this system.
}
\author{
Zhian N. Kamvar, Jonah Brooks
}

