% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{cromulent}
\alias{cromulent}
\title{Check for samples with incompatible missing data}
\usage{
cromulent(gid)
}
\arguments{
\item{gid}{a genind object}
}
\value{
a square matrix with samples that are incompatible
}
\description{
If two samples share no loci typed in common, they are incomparable and will
produce missing data in a distance matrix, which could lead to problems with
further analyses. This function finds these samples and returns a matrix of
how many other samples these are incompatible with.
}
\examples{

data(nancycats)
# These two populations have no incompatibilities
cromulent(nancycats[pop = c(1, 17)])
# If you reduce the number of loci, we find that there are incompatibilities
cromulent(nancycats[pop = c(1, 17), loc = c(1, 4)])
}

