% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstraping.R
\name{diversity_ci}
\alias{diversity_ci}
\title{Perform bootstrap statistics, calculate and plot confidence intervals.}
\usage{
diversity_ci(tab, n = 1000, ci = 95, total = TRUE, rarefy = FALSE,
  plot = TRUE, raw = TRUE, ...)
}
\arguments{
\item{tab}{a genind object OR a matrix produced from
\code{\link[poppr]{mlg.table}}.}

\item{n}{an integer defining the number of bootstrap replicates (defaults to
1000).}

\item{ci}{the percent for confidence interval.}

\item{total}{argument to be passed on to \code{\link[poppr]{mlg.table}} if
\code{tab} is a genind object.}

\item{rarefy}{if \code{TRUE}, bootstrapping will be performed on the smallest
population size. Defaults to \code{FALSE}, indicating that bootstrapping
will be performed respective to each population size.}

\item{plot}{If \code{TRUE} (default), boxplots will be produced for each
population, grouped by statistic. Colored dots will indicate the observed
value.This plot can be retrieved by using \code{p <- last_plot()} from the
\pkg{ggplot2} package.}

\item{raw}{if \code{TRUE} (default) a list containing three elements will be
returned}

\item{...}{parameters to be passed on to \code{\link[boot]{boot}} and
  \code{\link{diversity_stats}}}
}
\value{
\subsection{raw = TRUE}{
\itemize{
\item \strong{obs} - a matrix with observed statistics in columns, populations in rows
\item \strong{est} - a matrix with estimated statistics in columns, populations in rows
\item \strong{CI} - an array of 3 dimensions giving the lower and upper bound, the index measured, and the population.
\item \strong{boot} - a list containing the output of \code{\link[boot]{boot}} for each population.
  }
 }
\subsection{raw = FALSE}{ a data frame with the statistic observations,
estimates, and confidence intervals in columns, and populations in rows. Note
that the confidence intervals are converted to characters and rounded to
three decimal places. }
}
\description{
Perform bootstrap statistics, calculate and plot confidence intervals.
}
\details{
For details on the bootstrapping procedures, see
  \code{\link{diversity_boot}}. Confidence intervals are derived from the
  function \code{\link[boot]{norm.ci}}. The confidence interval is caluclated
  from the bootstrapped distribution and centered around the bias-corrected
  estimate. For rarefaction, the confidence interval is simply determined by
  caluclating the percentiles from the bootstrapped distribution.
}
\note{
While it is possible to use custom functions with this, there are three
  important things to remember when using these functions:
\enumerate{
\item The function must return a single value.
\item The function must allow for both matrix and vector inputs
\item The function name cannot match or partially match any arguments from
\code{\link[boot]{boot}}
} Anonymous functions are okay \cr(e.g. \code{function(x) vegan::rarefy(t(as.matrix(x)), 10)}).
}
\examples{
library(poppr)
data(Pinf)
diversity_ci(Pinf, n = 100L)
\dontrun{
# With pretty results
diversity_ci(Pinf, n = 100L, raw = FALSE)

# This can be done in a parallel fasion (OSX uses "multicore", Windows uses "snow")
system.time(diversity_ci(Pinf, 10000L, parallel = "multicore", ncpus = 4L))
system.time(diversity_ci(Pinf, 10000L))

# We often get many requests for a clonal fraction statistic. As this is
# simply the number of observed MLGs over the number of samples, we
# recommeneded that people calculate it themselves. With this function, you
# can add it in:

CF <- function(x){
 x <- drop(as.matrix(x))
 if (length(dim(x)) > 1){
   res <- rowSums(x > 0)/rowSums(x)
 } else {
   res <- sum(x > 0)/sum(x)
 }
 return(res)
}
# Show pretty results

diversity_ci(Pinf, 1000L, CF = CF, raw = FALSE)
diversity_ci(Pinf, 1000L, CF = CF, rarefy = TRUE, raw = FALSE)
}
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{diversity_boot}} \code{\link{diversity_stats}} \code{\link{poppr}} \code{\link[boot]{boot}} \code{\link[boot]{norm.ci}} \code{\link[boot]{boot.ci}}
}

