\docType{class}
\name{genclone-class}
\alias{genclone-class}
\title{genclone object}
\description{
An object based on the genind object that retains
multilocus genotype and population hierarchical
information.
}
\section{Slots}{

\describe{
\item{\code{mlg}}{a vector representing multilocus genotypes for the data set.}

\item{\code{hierarchy}}{a data frame containing hierarchical levels.}
}}
\section{Extends}{
  Class \code{"\linkS4class{genind}"}, directly.
}
\author{
Zhian N. Kamvar
}

