% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sandbox.r
\name{get_stats}
\alias{get_stats}
\title{Produce a table of diversity statistics}
\usage{
get_stats(z)
}
\arguments{
\item{z}{a table of integers representing counts of MLGs (columns) per
population (rows)}
}
\value{
a numeric matrix with 4 columns giving the following statistics for
  each population: \itemize{\item H - Shannon Diversity \item G Stoddart and
  Taylor's Diveristy (inverse Simpson's) \item unbiased Simpson Diversity \eqn{(N/(N-1))*(1 - D)} \item E.5 evenness}
}
\description{
Produce a table of diversity statistics
}
\examples{
library(poppr)
data(Pinf)
tab <- mlg.table(Pinf, bar = FALSE)
get_stats(tab)
}
\author{
Zhian N. Kamvar
}

