% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{is.snpclone}
\alias{is.snpclone}
\alias{is.clone}
\alias{is.genclone}
\title{Check for validity of a genclone or snpclone object}
\usage{
is.snpclone(x)

is.clone(x)

is.genclone(x)
}
\arguments{
\item{x}{a genclone or snpclone object}
}
\description{
Check for validity of a genclone or snpclone object
}
\note{
a \linkS4class{genclone} object will always be a valid 
\linkS4class{genind} object and a \linkS4class{snpclone} object will always
be a valid \linkS4class{genlight} object.
}
\examples{
(sc <- as.snpclone(glSim(100, 1e3, ploid=2, parallel = FALSE), 
                   parallel = FALSE, n.cores = 1L))
is.snpclone(sc)
is.clone(sc)
data(nancycats)
nanclone <- as.genclone(nancycats)
is.genclone(nanclone)
}
\author{
Zhian N. Kamvar
}
