\docType{methods}
\name{is.genclone}
\alias{is.genclone}
\alias{is.genclone,genclone-method}
\title{Check for validity of a genclone object}
\usage{
is.genclone(x)
}
\arguments{
  \item{x}{a genclone object}
}
\description{
In the hat
}
\examples{
data(nancycats)
nanclone <- as.genclone(nancycats)
is.genclone(nanclone)
}

