% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\name{is.snpclone}
\alias{is.snpclone}
\title{Check for validity of a snpclone object}
\usage{
is.snpclone(x)
}
\arguments{
\item{x}{a snpclone object}
}
\description{
Check for validity of a snpclone object
}
\note{
a \linkS4class{snpclone} object will always be a valid
\linkS4class{genlight} object.
}
\examples{
(x <- as.snpclone(glSim(100, 1e3, ploid=2)))
is.snpclone(x)
}
\author{
Zhian N. Kamvar
}

