% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R
\docType{methods}
\name{levels,MLG-method}
\alias{levels,MLG-method}
\alias{levels<-,MLG-method}
\title{Unique and Duplicated implementations for MLG objects}
\usage{
\S4method{levels}{MLG}(x)

\S4method{levels}{MLG}(x) <- value
}
\arguments{
\item{x}{an MLG object}
}
\value{
a character vector showing the levels of custom MLGs or NULL if the
visible slot is not set to "custom"
}
\description{
internal use
}
\keyword{internal}

