% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R
\name{visible}
\alias{MLG2df}
\alias{MLG2df,MLG-method}
\alias{visible}
\alias{visible,MLG-method}
\alias{visible<-}
\alias{visible<-,MLG-method}
\title{Accessors for the MLG object}
\usage{
visible(x)

visible(x) <- value

MLG2df(x)
}
\arguments{
\item{x}{an MLG object}

\item{value}{see details}
}
\value{
see details
}
\description{
Accessors for the MLG object
}
\examples{
\dontrun{
# These examples will simply show you what you can do with these
set.seed(5000)
(x <- sample(10, 20, replace = TRUE))
(m <- new("MLG", x))

# Visibility ------------------------------
visible(m) # original
visible(m) <- "custom"
m          # shows custom MLGS

# Conversion to data frame ----------------
MLG2df(m)  # Grab the internal data frame

}
}
\keyword{internal}

