% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R, R/methods.r
\docType{methods}
\name{initialize,MLG-method}
\alias{MLG}
\alias{MLG,genclone-method}
\alias{MLG<-}
\alias{MLG<-,genclone-method}
\alias{Math,MLG-method}
\alias{Math2,MLG-method}
\alias{Ops,MLG,ANY-method}
\alias{Summary,MLG-method}
\alias{[,MLG,ANY,ANY,ANY-method}
\alias{[<-,MLG,ANY,ANY,ANY-method}
\alias{initialize,MLG-method}
\alias{length,MLG-method}
\alias{show,MLG-method}
\title{Methods used for MLG objects}
\usage{
\S4method{initialize}{MLG}(.Object, mlg)

\S4method{[}{MLG,ANY,ANY,ANY}(x, i, j, ..., all = FALSE, drop = TRUE)

\S4method{[}{MLG,ANY,ANY,ANY}(x, i, j) <- value

\S4method{show}{MLG}(object)

\S4method{length}{MLG}(x)

\S4method{Ops}{MLG,ANY}(e1, e2)

\S4method{Math}{MLG}(x)

\S4method{Math2}{MLG}(x, digits)

\S4method{Summary}{MLG}(x, ..., na.rm = FALSE)

MLG(x, type = NULL)

MLG(x) <- value
}
\arguments{
\item{.Object}{a character, "MLG"}

\item{mlg}{a vector where each element assigns the multilocus genotype of
that individual in the data set.}

\item{x}{an MLG object}

\item{i}{a vector of integers or logical values to index the MLG vector.}

\item{j}{One of
\itemize{
 \item "original"
 \item "expanded"
 \item "contracted"
 \item "custom"
}}

\item{...}{passed on to summary methods}

\item{all}{a logical value indicating whether or not to return the subset of
all MLG values or only the numeric.}

\item{value}{the vector of MLGs to replace. \strong{For internal use only}.}

\item{object}{an MLG object.}

\item{e1}{an MLG object}

\item{e2}{a number}

\item{digits}{the number of digits to retain}

\item{na.rm}{passed on to summary methods}

\item{type}{a character specifying "original", "contracted", or "custom"
defining they type of mlgs to return. Defaults to what is set in the
object.}

\item{x}{a \linkS4class{genclone} or \linkS4class{snpclone} object.}

\item{value}{a character specifying which mlg type is visible in the object.
  See details.}
}
\value{
an object of the same type as x.
}
\description{
Default methods for acessing and subsetting MLG objects.

The following methods allow the user to access and manipulate multilocus
genotypes in genclone or snpclone objects.
}
\details{
\linkS4class{genclone} and \linkS4class{snpclone} objects have a
  slot for an internal class of object called \linkS4class{MLG}. This class
  allows the storage of flexible MLG definitions: \itemize{ \item "original"
  - naive mlgs defined by string comparison. This is default. \item
  "contracted" - mlgs defined by a genetic distance threshold. \item "custom"
  - user-defined MLGs }
}
\examples{
data(partial_clone)
pc <- as.genclone(partial_clone)
MLG(pc)
MLG(pc) <- "custom"
MLG(pc)
mlg.levels(pc) <- LETTERS
MLG(pc)
}
\author{
Zhian N. Kamvar

Zhian N. Kamvar
}
\keyword{internal}

