% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.r
\name{mlg.crossplot}
\alias{mlg.crossplot}
\title{Create a circular plot to show MLGs crossing populations [EXPERIMENTAL]}
\usage{
mlg.crossplot(gid, pop = NULL, temporal = NULL, PAL = grey.colors)
}
\arguments{
\item{gid}{a genind/genclone/genlight/snpclone object}

\item{pop}{an optional population factor or formula to be passed to \code{\link{setPop}}}

\item{temporal}{a character specifying the temporal strata.}

\item{PAL}{a color palette to use for coloring edges.}
}
\value{
an igraph object
}
\description{
Create a circular plot to show MLGs crossing populations [EXPERIMENTAL]
}
\examples{
data(Pram)
x <- mlg.crossplot(Pram, ~SOURCE/STATE, ~YEAR)
plot(x)
\dontrun{
# You can use the viridis palette
if (require("viridis")){
  plot(mlg.crossplot(Pram, ~SOURCE/STATE, ~YEAR, PAL = viridis_pal()))
}
}
}
\author{
Zhian N. Kamvar
}

