% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{mlg.custom}
\alias{mlg.custom}
\alias{mlg.custom,genclone-method}
\alias{mlg.custom<-}
\alias{mlg.custom<-,genclone-method}
\alias{mlg.levels}
\alias{mlg.levels,genclone-method}
\alias{mlg.levels<-}
\alias{mlg.levels<-,genclone-method}
\title{Define custom multilocus genotypes}
\usage{
mlg.custom(x, set = TRUE, value)

mlg.custom(x, set = TRUE) <- value

mlg.levels(x, set = TRUE, value)

mlg.levels(x, set = TRUE) <- value
}
\arguments{
\item{x}{a \linkS4class{genclone} or \linkS4class{snpclone} object.}

\item{set}{logical. If \code{TRUE} (default), the visible MLGs will be set to
'custom'.}

\item{value}{a vector that defines the multilocus genotypes for your data.
This can be a vector of ANYTHING that can be turned into a factor.}
}
\value{
an object of the same type as x
}
\description{
This function will allow you to define custom multilocus genotypes for your
data set.
}
\examples{
data(partial_clone)
pc <- as.genclone(partial_clone)
mlg.custom(pc) <- LETTERS[MLG(pc)]
MLG(pc)
MLG(pc) <- "original"
MLG(pc)
}
\author{
Zhian N. Kamvar
}

