% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R
\docType{methods}
\name{setmlg}
\alias{setmlg}
\alias{setmlg,MLG-method}
\alias{setmlg<-}
\alias{setmlg<-,MLG-method}
\title{Access and manipluate multilocus genotypes or lineages}
\usage{
setmlg(x, value = NULL)

setmlg(x) <- value
}
\arguments{
\item{x}{an MLG object}

\item{value}{one of the 4 possible MLG types.}
}
\description{
The MLG object in a \linkS4class{genclone} object contains 4 vectors to give
the user flexibility of MLG assignment.
}
\details{
The four possible MLG types:

\itemize{
\item "original" - Strict definition of MLGs. Sensitive to small changes in
data and missing data.
\item "contracted" - aka Multilocus Lineages. These MLGs are collapsed into
groups given a certain genetic distance, cutoff value, and algorithm
\item "expanded" - Expanded multilocus genotype groups for when there are too
few loci. Psex is caluclated and mlgs are split into new MLGs based on a given
pvalue cutoff of Psex.
\item "custom" - User defined MLGs
}
}

