% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\name{old2new_genclone}
\alias{old2new_genclone}
\title{Convert an old genclone object to a new genclone object}
\usage{
old2new_genclone(object, donor = new(class(object)))
}
\arguments{
\item{object}{a genclone object from poppr v. 1.1}

\item{donor}{a new genclone object from poppr v. 2.0}
}
\description{
Convert an old genclone object to a new genclone object
}
\author{
Zhian N. Kamvar
}

