% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_robin.R
\name{pgen}
\alias{pgen}
\title{Genotype Probability}
\usage{
pgen(x, log = TRUE, by_pop = TRUE, freq = NULL)
}
\arguments{
\item{x}{a genind or genclone object.}

\item{log}{a \code{logical} if \code{log =TRUE} (default), the values
  returned will be log(Pgen). If \code{log = FALSE}, the values returned will
  be Pgen.}

\item{by_pop}{a \code{logical} to determine whether allelic frequencies
  should be calculated per population (\code{TRUE}, default) or across all
  populations in the data (\code{FALSE}).}

\item{freq}{a vector or matrix of allele frequencies. This defaults to
  \code{NULL}, indicating that the frequencies will be determined via
  round-robin approach.}
}
\value{
A vector containing Pgen values per locus for each genotype in the
  object.
}
\description{
Genotype Probability
}
\note{
For haploids, Pgen at a particular locus is the allele frequency.
}
\examples{
data(Pram)
head(pgen(Pram, log = FALSE))

\dontrun{
# You can get the Pgen values over all loci by summing over the logged results:
exp(rowSums(pgen(Pram, log = TRUE, na.rm = TRUE)))

# You can also take the product of the non-logged results:
apply(pgen(Pram, log = FALSE), 1, prod, na.rm = TRUE)
}
}
\author{
Zhian N. Kamvar, Jonah Brooks, Stacy Hadfield
}
\seealso{
\code{\link{psex}}, \code{\link{rraf}}, \code{\link{rrmlg}}
}

