\docType{data}
\name{Pinf}
\alias{Pinf}
\title{Phytophthora infestans data from Mexico and South America.}
\format{a \code{\linkS4class{genclone}} object with 2 population hierarchies
  called "Continent" and "Country" that contain 2 and 4 populations,
  respectively.}
\usage{
data(Pinf)
}
\description{
The Pinf data set contains 86 isolates genotyped over 11
microsatellite loci collected from Mexico, Peru, Columbia,
and Ecuador.  This is a subset of the data used for the
reference below.
}
\references{
Goss, EM et al. submitted, 2014.
}

