% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{poppr-package}
\alias{poppr-package}
\title{Tools for analysys of populations with clonal, partially clonal, and/or
sexual reproduction.}
\description{
Poppr provides tools for population genetic analysis that include genotypic
diveristy measures, genetic distances with bootstrap support, native
organization and handling of population hierarchies, and clone correction.
}
\details{
This package documentation is in progress.
}
\references{
Kamvar ZN, Tabima JF, Grünwald NJ. (2014) Poppr: an R package for
genetic analysis of populations with clonal, partially clonal, and/or sexual
reproduction. PeerJ 2:e281 http://dx.doi.org/10.7717/peerj.281
}

