\docType{methods}
\name{setpop}
\alias{setpop}
\alias{setpop<-}
\title{Manipulate the population factor of genclone objects.}
\usage{
setpop(x, formula = NULL)

setpop(x) <- value
}
\arguments{
  \item{x}{a genclone object}

  \item{formula}{a nested formula indicating the order of
  the population hierarchy.}

  \item{value}{same as formula}
}
\description{
The following methods allow the user to quickly change the
population of a genclone object.
}

