% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_robin.R
\name{psex}
\alias{psex}
\title{Probability of encountering a genotype more than once by chance}
\usage{
psex(x, by_pop = TRUE)
}
\arguments{
\item{x}{a genind or genclone object.}

\item{by_pop}{a \code{logical} to determine whether allelic frequencies
should be calculated per population (\code{TRUE}, default) or across all
populations in the data (\code{FALSE}).}
}
\value{
a vector of Psex for each sample.
}
\description{
Probability of encountering a genotype more than once by chance
}
\note{
The values of Psex represent the value for each multilocus genotype.
}
\examples{
data(Pram)
Pram_psex <- psex(Pram, by_pop = FALSE)
plot(Pram_psex, log = "y", col = ifelse(Pram_psex > 0.05, "red", "blue"))
abline(h = 0.05, lty = 2)
\dontrun{
# This can be also done assuming populations structure
Pram_psex <- psex(Pram, by_pop = TRUE)
plot(Pram_psex, log = "y", col = ifelse(Pram_psex > 0.05, "red", "blue"))
abline(h = 0.05, lty = 2)
}
}
\author{
Zhian N. Kamvar, Jonah Brooks, Stacy Hadfield
}
\seealso{
\code{\link{pgen}}, \code{\link{rraf}}, \code{\link{rrmlg}}
}

