% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{as.snpclone}
\alias{as.snpclone}
\alias{as.snpclone,genlight-method}
\title{Create a snpclone object from a genlight object.}
\usage{
as.snpclone(x, hierarchy = NULL)
}
\arguments{
\item{x}{a \code{\linkS4class{genlight}} or \code{\linkS4class{snpclone}}
object}

\item{hierarchy}{a data frame representing the population hierarchy.}
}
\description{
Wrapper for snpclone initializer.
}
\note{
The hierarchy must have the same number of rows as the number of
  observations in the genlight object. If no hierarchy is defined, the function
  will search for a data frame in the \code{\link{other}} slot called
  "population_hierarchy" and set that as the hieararchy. If none is defined,
  the population will be set as the hierarchy under the label "Pop". Use the
  function \code{\link{splithierarchy}} to split up any population
  hierarchies that might be combined in the population factor.
}
\examples{
(x <- as.snpclone(glSim(100, 1e3, ploid=2)))
}
\author{
Zhian N. Kamvar
}

