% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{[,snpclone,ANY,ANY,ANY-method}
\alias{[,snpclone,ANY,ANY,ANY-method}
\alias{initialize,snpclone-method}
\alias{show,snpclone-method}
\title{Methods used for the snpclone object}
\usage{
\S4method{[}{snpclone,ANY,ANY,ANY}(x, i, j, ..., mlg.reset = FALSE, drop = FALSE)

\S4method{initialize}{snpclone}(.Object, ..., mlg, mlgclass = TRUE)

\S4method{show}{snpclone}(object)
}
\arguments{
\item{x}{a snpclone object}

\item{i}{vector of numerics indicating number of individuals desired}

\item{j}{a vector of numerics corresponding to the loci desired.}

\item{...}{passed on to the \code{\linkS4class{genlight}} object.}

\item{mlg.reset}{logical. Defaults to \code{FALSE}. If \code{TRUE}, the mlg
vector will be reset}

\item{drop}{set to \code{FALSE}}

\item{.Object}{a character, "snpclone"}

\item{mlg}{a vector where each element assigns the multilocus genotype of
that individual in the data set.}

\item{mlgclass}{a logical value specifying whether or not to translate the
mlg object into an MLG class object.}

\item{object}{a snpclone object}
}
\description{
Default methods for subsetting snpclone objects.
}
\author{
Zhian N. Kamvar

Zhian N. Kamvar
}
\keyword{internal}
