% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R
\docType{methods}
\name{unique.MLG}
\alias{duplicated,MLG-method}
\alias{duplicated.MLG}
\alias{unique,MLG-method}
\alias{unique.MLG}
\title{Unique and Duplicated implementations for MLG objects}
\usage{
unique.MLG(x, incomparables = FALSE, ...)

duplicated.MLG(x, incomparables = FALSE, ...)
}
\arguments{
\item{incomparables}{vector of values that cannot be compared}

\item{...}{options passed on to the base function \link{unique} or \link{duplicated}.}
}
\description{
internal use
}
\keyword{internal}

