% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgma.R
\name{upgma}
\alias{upgma}
\title{UPGMA}
\usage{
upgma(d)
}
\arguments{
\item{d}{A distance matrix.}
}
\value{
A phylogenetic tree of class \code{phylo}.
}
\description{
UPGMA clustering. Just a wrapper function around \code{\link[stats]{hclust}}.
}
\examples{

library(ape)
data(woodmouse)
dm <- dist.dna(woodmouse)
tree <- upgma(dm)
plot(tree)

}
\seealso{
\code{\link{hclust}}, \code{\link[ape]{as.phylo}}
}
\author{
Klaus Schliep \email{klaus.schliep@gmail.com}
}
\keyword{cluster}
