\docType{data}
\name{Aeut}
\alias{Aeut}
\title{Oomycete root rot pathogen \emph{Aphanomyces euteiches} AFLP data}
\format{a \code{\link{genind}} object with two populations containing a data
  frame in the \code{other} slot called \code{population_hierarchy}. This
  data frame gives indices of the populations and subpopulations for the data
  set.}
\usage{
data(Aeut)
}
\description{
The Aeut dataset consists of 187 isolates of the Oomycete
root rot pathogen, \emph{Aphanomyces euteiches} collected
from two different fields in NW Oregon and W Washington,
USA.
}
\references{
Grunwald, NJ and Hoheisel, G.A. 2006. Hierarchical Analysis
of Diversity, Selfing, and Genetic Differentiation in
Populations of the Oomycete \emph{Aphanomyces euteiches}.
Phytopathology 96:1134-1141
}

