% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bitwise.dist}
\alias{bitwise.dist}
\title{Calculate a distance matrix comparing samples based on the number of alleles
that differ in zygosity.}
\usage{
bitwise.dist(x, diff = TRUE, percent = TRUE, mat = FALSE)
}
\arguments{
\item{x}{a genlight object.}

\item{diff}{\code{logical}. If \code{TRUE}, this will count the number of
differences between individuals. \code{FALSE} will count the number of
similarities. Default set to \code{TRUE}}

\item{percent}{\code{logical}. Should the distance be represented from 0 to 1?
Default set to \code{TRUE}. \code{FALSE} will return the distance represented
as integers from 1 to n where n is the number of loci.}

\item{mat}{\code{logical}. Return a matrix object. Default set to
\code{FALSE}, returning a dist object. \code{TRUE} returns a matrix object.}
}
\value{
Pairwise distances between individuals present in the genlight object.
}
\description{
This function does pairwise comparisons between diploid samples in a genlight
object. The number representing the distance between two samples is equal to
the number of alleles in the samples that do not have the same zygosity.
}
\author{
Zhian N. Kamvar, Jonah Brooks
}

