\docType{class}
\name{bootgen-class}
\alias{bootgen-class}
\title{Bootgen object}
\description{
An internal object used for bootstrapping. Not intended for
user interaction.
}
\section{Slots}{

\describe{
\item{\code{replen}}{repeat length of microsatellite loci}
}}
\section{Extends}{
  Class \code{"\linkS4class{genind}"}, directly.
}
\author{
Zhian Kamvar
}

