\docType{class}
\name{bootgen-class}
\alias{bootgen-class}
\title{Bootgen object}
\description{
An internal object used for bootstrapping. Not intended for
user interaction.
}
\section{Slots}{

\describe{
\item{\code{alllist}}{a list with numeric vectors, each representing a different
  locus where each element in the vector represents the index for a specific
  allele.}
}}
\section{Extends}{
  Class \code{"\linkS4class{genclone}"}, directly.
}
\author{
Zhian N. Kamvar
}

