% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{[,bootgen,ANY,ANY,ANY-method}
\alias{$,bootgen-method}
\alias{[,bootgen,ANY,ANY,ANY-method}
\alias{dim,bootgen-method}
\alias{initialize,bootgen-method}
\title{Methods used for the bootgen object.}
\usage{
\S4method{[}{bootgen,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{dim}{bootgen}(x)

\S4method{$}{bootgen}(x, name)

\S4method{initialize}{bootgen}(.Object, gen)
}
\arguments{
\item{x}{a \code{"\linkS4class{bootgen}"} object}

\item{i}{vector of numerics indicating number of individuals desired}

\item{j}{a vector of numerics corresponding to the loci desired.}

\item{...}{unused.}

\item{drop}{set to \code{FALSE}}

\item{.Object}{a character, "bootgen"}

\item{gen}{a genind, genclone, or genpop object}

\item{name}{
    A literal character string or a \link{name} (possibly \link{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link{names}}
    of the object.
  }
}
\description{
This is not designed for user interaction.
}
\author{
Zhian N. Kamvar
}
\keyword{internal}

