% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,bootgen,ANY,ANY,ANY-method}
\alias{$,bootgen-method}
\alias{[,bootgen,ANY,ANY,ANY-method}
\alias{dim,bootgen-method}
\alias{initialize,bootgen-method}
\title{Methods used for the bootgen object.}
\usage{
\S4method{[}{bootgen,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{dim}{bootgen}(x)

\S4method{$}{bootgen}(x, name)

\S4method{initialize}{bootgen}(.Object, gen, na = "mean", freq = TRUE)
}
\arguments{
\item{x}{a \code{"\linkS4class{bootgen}"} object}

\item{i}{vector of numerics indicating number of individuals desired}

\item{j}{a vector of numerics corresponding to the loci desired.}

\item{...}{unused.}

\item{drop}{set to \code{FALSE}}

\item{.Object}{a character, "bootgen"}

\item{gen}{a genind, genclone, or genpop object}

\item{na}{how missing data should be treated. Default is "mean", averaging
over other values in the matrix. Possible values are listed in
\code{\link[adegenet]{tab}}.}

\item{freq}{if \code{TRUE}, the matrix will be a genotype frequency matrix.
If \code{FALSE}, the matrix will be allele counts.}
}
\description{
This is not designed for user interaction.
}
\author{
Zhian N. Kamvar
}
\keyword{internal}

