% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{initialize,bruvomat-method}
\alias{initialize,bruvomat-method}
\alias{dim,bruvomat-method}
\alias{[,bruvomat,ANY,ANY,ANY-method}
\title{Methods used for the bruvomat object.}
\usage{
\S4method{initialize}{bruvomat}(.Object, gen, replen)

\S4method{dim}{bruvomat}(x)

\S4method{[}{bruvomat,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{.Object}{a character, "bruvomat"}

\item{gen}{\code{"\linkS4class{genind}"} object}

\item{replen}{a vector of numbers indicating the repeat length for each 
microsatellite locus.}

\item{x}{a \code{"\linkS4class{bruvomat}"} object}

\item{i}{vector of numerics indicating number of individuals desired}

\item{j}{a vector of numerics corresponding to the loci desired.}

\item{...}{unused.}

\item{drop}{set to \code{FALSE}}
}
\description{
This is not designed for user interaction.
}
\author{
Zhian N. Kamvar
}
\keyword{internal}
