% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{as.genclone}
\alias{as.genclone}
\alias{as.genclone,genind-method}
\title{Create a genclone object from a genind object.}
\usage{
as.genclone(x, ..., mlg, mlgclass = TRUE)
}
\arguments{
\item{x}{a \code{\linkS4class{genind}} or \code{\linkS4class{genclone}}
object}

\item{...}{arguments passed on to the \code{\linkS4class{genind}} constructor}

\item{mlg}{an optional vector of multilocus genotypes as integers}

\item{mlgclass}{should the mlg slot be of class MLG?}
}
\description{
Wrapper for genclone initializer.
}
\examples{
data(Aeut)
Aeut
Aeut.gc <- as.genclone(Aeut)
Aeut.gc
Aeut.gc <- as.genclone(Aeut)
Aeut.gc
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{splitStrata}}, \code{\linkS4class{genclone}},
  \code{\link{read.genalex}}
}

