% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gethierarchy}
\alias{addhierarchy}
\alias{addhierarchy<-}
\alias{gethierarchy}
\alias{namehierarchy}
\alias{namehierarchy<-}
\alias{sethierarchy}
\alias{sethierarchy<-}
\alias{splithierarchy}
\alias{splithierarchy<-}
\title{DEPRECATED FUNCTIONS}
\usage{
gethierarchy(...)

sethierarchy(...)

sethierarchy(...) <- value

addhierarchy(...)

addhierarchy(...) <- value

splithierarchy(...)

splithierarchy(...) <- value

namehierarchy(...)

namehierarchy(...) <- value
}
\arguments{
\item{...}{arguments to be passed on to adegenet
\code{\link[adegenet]{strata}} methods.}

\item{value}{argument passed on to adegenet methods.}
}
\value{
a deprecation message and the same return as the adegenet methods.
}
\description{
These functions are deprecated as of \pkg{poppr} version 2.0 and have been
moved to the \pkg{adegenet} package. See \code{\link[adegenet]{strata}} for
details.
}
\section{Background}{

  These functions origianlly appeared in \pkg{poppr} in July of 2014. In
  March 2015, a request was made for them to be ported to the \pkg{adegenet}
  package as part of the population genetics in R hackathon. The name of the
  functions were moved from "hierarchy" to "strata" because it better
  relfected the nature of the fields to contain any information possible.
}
\seealso{
\code{\link[adegenet]{strata}}
}

