% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gethierarchy}
\alias{addhierarchy}
\alias{addhierarchy<-}
\alias{gethierarchy}
\alias{namehierarchy}
\alias{namehierarchy<-}
\alias{sethierarchy}
\alias{sethierarchy<-}
\alias{splithierarchy}
\alias{splithierarchy<-}
\title{DEPRECATED FUNCTIONS}
\usage{
gethierarchy(...)

sethierarchy(...)

sethierarchy(...) <- value

addhierarchy(...)

addhierarchy(...) <- value

splithierarchy(...)

splithierarchy(...) <- value

namehierarchy(...)

namehierarchy(...) <- value
}
\arguments{
\item{...}{arguments to be passed on to adegenet \code{\link[adegenet]{strata}} methods.}
}
\value{
a deprecation message and the same return as the adegenet methods.
}
\description{
These functions are deprecated as of \pkg{poppr} version 2.0 and have been
moved to the \pkg{adegenet} package. See \code{\link[adegenet]{strata}} for
details.
}
\seealso{
\code{\link[adegenet]{strata}}
}

