% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstraping.R
\name{diversity_boot}
\alias{diversity_boot}
\title{Perform a bootstrap analysis on diversity statistics}
\usage{
diversity_boot(tab, n, n.rare = NULL, mlg.weight = TRUE, H = TRUE,
  G = TRUE, lambda = TRUE, E5 = TRUE, ...)
}
\arguments{
\item{tab}{a table produced from the \pkg{poppr} function
\code{\link[poppr]{mlg.table}}. MLGs in columns and populations in rows}

\item{n}{an integer > 0 specifying the number of bootstrap replicates to
perform (corresponds to \code{R} in the function \code{\link[boot]{boot}}.}

\item{n.rare}{a sample size at which all resamplings should be performed.
This should be no larger than the smallest sample size. Defaults to
\code{NULL}, indicating that each population will be sampled at its own
size.}

\item{mlg.weight}{when \code{FALSE}, all observed MLGs have an equal chance
of being selected in the bootstrapping procedure. When \code{TRUE}
(default), larger MLGs have more of a chance of being selected. This
parameter is ignored when \code{n.rare = TRUE}.}

\item{H}{logical whether or not to calculate Shannon's index}

\item{G}{logical whether or not to calculate Stoddart and Taylor's index}

\item{lambda}{logical whether or not to calculate Simpson's index}

\item{E5}{logical whether or not to calculate Evenness}

\item{...}{other parameters passed on to \code{\link[boot]{boot}} and
  \code{\link{diversity_stats}}.}
}
\value{
a list of objects of class "boot".
}
\description{
Perform a bootstrap analysis on diversity statistics
}
\examples{
library(poppr)
data(Pinf)
tab <- mlg.table(Pinf, plot = FALSE)
diversity_boot(tab, 10L)
\dontrun{
# This can be done in a parallel fasion (OSX uses "multicore", Windows uses "snow")
system.time(diversity_boot(tab, 10000L, parallel = "multicore", ncpus = 4L))
system.time(diversity_boot(tab, 10000L))
}
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{diversity_stats}} for basic statistic calculation,
   \code{\link{diversity_ci}} for confidence intervals and plotting, and
   \code{\link{poppr}}
}

