% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{methods}
\name{[,genclone,ANY,ANY,ANY-method}
\alias{[,genclone,ANY,ANY,ANY-method}
\alias{initialize,genclone-method}
\alias{print,genclone-method}
\alias{show,genclone-method}
\title{Methods used for the genclone object}
\usage{
\S4method{[}{genclone,ANY,ANY,ANY}(x, i, j, ..., loc = NULL,
  treatOther = TRUE, quiet = TRUE, drop = FALSE)

\S4method{initialize}{genclone}(.Object, gen, hierarchy, mlg)

\S4method{show}{genclone}(object)

\S4method{print}{genclone}(x, ...)
}
\arguments{
\item{x}{a genclone object}

\item{i}{vector of numerics indicating number of individuals desired}

\item{j}{a vector of numerics corresponding to the loci desired.}

\item{...}{passed on to the \code{\linkS4class{genind}} object.}

\item{drop}{set to \code{FALSE}}

\item{loc}{passed on to \code{\linkS4class{genind}} object.}

\item{treatOther}{passed on to \code{\linkS4class{genind}} object.}

\item{quiet}{passed on to \code{\linkS4class{genind}} object.}

\item{.Object}{a character, "genclone"}

\item{gen}{\code{"\linkS4class{genind}"} object}

\item{hierarchy}{a data frame where each row i represents the different
population assignments of individual i in the data set. If this is empty, the
hierarchy will be created from the population factor.}

\item{mlg}{a vector where each element assigns the multilocus genotype of
that individual in the data set.}

\item{object}{a genclone object}

\item{x}{a genclone object}

\item{fullnames}{\code{logical}. If \code{TRUE}, then the full names of the
populations will be printed. If \code{FALSE}, then only the first and last
three population names are displayed.}
}
\description{
Default methods for subsetting genclone objects.
}
\author{
Zhian N. Kamvar
}
\keyword{internal}

