% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.r
\name{genclone.update}
\alias{genclone.update}
\title{Updates old genclone objects to the new format}
\usage{
genclone.update(x)
}
\arguments{
\item{x}{a genclone object}
}
\description{
Returns an updated genclone object containing a default mll slot.
}
\examples{
data(Pinf)
Pinf <- genclone.update(Pinf)
Pinf$mll
}
\author{
Jonah C. Brooks
}

