\docType{methods}
\name{hierarchy}
\alias{hierarchy}
\alias{hierarchy,genclone-method}
\title{Access and organize the population hierarchy for genclone objects.}
\usage{
hierarchy(x, ...)

\S4method{hierarchy}{genclone}(x, ...)
}
\arguments{
  \item{x}{a genclone object}

  \item{...}{I'll fiddle with this later.}
}
\description{
word
}

