\docType{methods}
\name{hierarchy}
\alias{hierarchy}
\alias{hierarchy,genclone-method}
\alias{hierarchy<-}
\alias{hierarchy<-,genclone-method}
\title{Access and organize the population hierarchy for genclone objects.}
\usage{
hierarchy(x, ...)

\S4method{hierarchy}{genclone}(x, ...)

hierarchy(x) <- value

\S4method{hierarchy}{genclone}(x) <- value
}
\arguments{
  \item{x}{a genclone object}

  \item{...}{I'll fiddle with this later.}

  \item{value}{a data frame to replace the population
  hierarchy with.}
}
\description{
word
}

