% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\name{is.genclone}
\alias{is.genclone}
\title{Check for validity of a genclone object}
\usage{
is.genclone(x)
}
\arguments{
\item{x}{a genclone object}
}
\description{
Check for validity of a genclone object
}
\note{
a \linkS4class{genclone} object will always be a valid
\linkS4class{genind} object.
}
\examples{
data(nancycats)
nanclone <- as.genclone(nancycats)
is.genclone(nanclone)
}
\author{
Zhian N. Kamvar
}

