% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{is.genclone}
\alias{is.genclone}
\title{Check for validity of a genclone object}
\usage{
is.genclone(x)
}
\arguments{
\item{x}{a genclone object}
}
\description{
Note that a \linkS4class{genclone} object will always be a valid
\linkS4class{genind} object.
}
\examples{
data(nancycats)
nanclone <- as.genclone(nancycats)
is.genclone(nanclone)
}
\author{
Zhian N. Kamvar
}

