% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_robin.R
\name{minor_allele_correction}
\alias{minor_allele_correction}
\title{Correct minor allele frequencies derived from rraf}
\usage{
minor_allele_correction(rraf, rrmlg, mlg = NULL, pop = NULL,
  locfac = NULL, e = NULL, sum_to_one = FALSE, d = c("sample", "mlg",
  "rrmlg"), m = 1)
}
\arguments{
\item{rraf}{a list or matrix produced from rraf (with \code{correction = 
FALSE})}

\item{rrmlg}{a matrix containing multilocus genotypes per locus}

\item{mlg}{the number of MLGs in the sample. Only required if \code{d = 
"mlg"}.}

\item{pop}{a vector of factors that define the population definition for each
observation in \code{rrmlg}. This must be supplied if \code{rraf} is a 
matrix.}

\item{locfac}{a vector of factors that define the columns belonging to the 
loci.}

\item{e}{a numeric epsilon value to use for all missing allele frequencies.}

\item{sum_to_one}{when \code{TRUE}, the original frequencies will be reduced 
so that all allele frequencies will sum to one. \strong{Default: 
\code{FALSE}}}

\item{d}{the unit by which to take the reciprocal. \code{div = "sample"} will
be 1/(n samples), \code{div = "mlg"} will be 1/(n mlg), and \code{div = 
"rrmlg"} will be 1/(n mlg at that locus). This is overridden by \code{e}.}

\item{m}{a multiplier for div. Default is \code{mult = 1}. This parameter is
overridden by \code{e}}
}
\value{
a matrix or vector the same type as rraf
}
\description{
Correct minor allele frequencies derived from rraf
}
\details{
By default, this will add 1/(n samples) to all zero-value alleles.
}
\examples{
data(Pram)
prrmlg <- rrmlg(Pram)
pramaf <- rraf(Pram, correction = FALSE)
minor_allele_correction(pramaf, prrmlg)
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{rraf}},
  \code{\link{pgen}},
  \code{\link{psex}}
}
\keyword{internal}

