% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaaMLGclass.R
\docType{class}
\name{MLG-class}
\alias{MLG}
\alias{MLG-class}
\title{MLG class}
\description{
A class to store multilocus genotypes in genclone objects.
}
\section{Slots}{

\describe{
\item{\code{mlg}}{a list containing four vectors, one for each type of MLG manipulation.}

\item{\code{visible}}{a character specifying which MLG type is to be displayed and accessed.}

\item{\code{dist}}{a function to calculate the genetic distance for collapsing MLGs.}

\item{\code{distname}}{the name of the distance function}

\item{\code{Two}}{numbers specifying the cutoff value for expanding and collapsing
MLGs.}
}}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{as.genclone}} \code{\link{sethierarchy}} \code{\link{setpop}}
\code{\linkS4class{genind}}
}

