% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_robin.R
\name{pgen}
\alias{pgen}
\title{Genotype Probability}
\usage{
pgen(gid, pop = NULL, by_pop = TRUE, log = TRUE, freq = NULL)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{pop}{either a formula to set the population factor from the
\code{\link{strata}} slot or a vector specifying the population factor for
each sample. Defaults to NULL.}

\item{by_pop}{When this is \code{TRUE}, the calculation will be done by
population. Defaults to \code{FALSE}}

\item{log}{a \code{logical} if \code{log =TRUE} (default), the values
  returned will be log(Pgen). If \code{log = FALSE}, the values returned will
  be Pgen.}

\item{freq}{a vector or matrix of allele frequencies. This defaults to
  \code{NULL}, indicating that the frequencies will be determined via
  round-robin approach.}

\item{x}{a genind or genclone object.}
}
\value{
A vector containing Pgen values per locus for each genotype in the
  object.
}
\description{
Genotype Probability
}
\note{
For haploids, Pgen at a particular locus is the allele frequency. This
  function cannot handle polyploids. Additionally, when the argument
  \code{pop} is not \code{NULL}, \code{by_pop} is automatically \code{TRUE}.
}
\examples{
data(Pram)
head(pgen(Pram, log = FALSE))

\dontrun{
# You can get the Pgen values over all loci by summing over the logged results:
exp(rowSums(pgen(Pram, log = TRUE, na.rm = TRUE)))

# You can also take the product of the non-logged results:
apply(pgen(Pram, log = FALSE), 1, prod, na.rm = TRUE)
}
}
\author{
Zhian N. Kamvar, Jonah Brooks, Stacy Hadfield
}
\seealso{
\code{\link{psex}}, \code{\link{rraf}}, \code{\link{rrmlg}}
}

