% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_stats.R
\name{plot_filter_stats}
\alias{plot_filter_stats}
\title{Plot the results of filter_stats}
\usage{
plot_filter_stats(x, fstats, distmat, cols = NULL, nclone = NULL,
  breaks = NULL)
}
\arguments{
\item{x}{a genlight of genind object}

\item{fstats}{the list passed from \code{\link{filter_stats}}}

\item{distmat}{a distance matrix passed from \code{\link{filter_stats}}}

\item{cols}{colors to use for each algorithm (defaults to \pkg{RColorBrewer}
set 1)}

\item{nclone}{see \code{\link{filter_stats}}}
}
\value{
a plot depicting how many MLLs are collapsed as the genetic distance
  increases for each algorithm.
}
\description{
Plot the results of filter_stats
}
\note{
This function originally appeared in
  \href{http://dx.doi.org/10.5281/zenodo.17424}{DOI: 10.5281/zenodo.17424}
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary
Material for Frontiers Plant Genetics and Genomics 'Novel R tools for
analysis of genome-wide population genetic data with emphasis on clonality'.
DOI: \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}

Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of
genome-wide population genetic data with emphasis on clonality. Front. Genet.
6:208. doi:
\href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}
\keyword{internal}

