% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwise.r
\name{poppr_has_parallel}
\alias{poppr_has_parallel}
\title{Determines whether openMP is support on this system.}
\usage{
poppr_has_parallel()
}
\value{
FALSE if openMP is not supported, TRUE if it is
}
\description{
Determines whether openMP is support on this system.
}
\examples{
poppr_has_parallel()
}
\author{
Zhian N. Kamvar, Jonah C. Brooks
}
