% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_robin.R
\name{rraf}
\alias{rraf}
\title{Round Robin Allele Frequencies}
\usage{
rraf(gid, res = "list")
}
\arguments{
\item{gid}{a genind or genclone object}

\item{res}{Either "list" (default) or "vector".}
}
\value{
a vector or list of allele frequencies
}
\description{
This function utilizes \code{\link{rrmlg}} to calculate multilocus genotypes
and then subsets each locus by the resulting MLGs to calculate the
round-robin allele frequencies used for pgen and psex.
}
\author{
Zhian N. Kamvar, Jonah C. Brooks, Stacey Hatfield
}

