% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round_robin.R
\name{rraf}
\alias{rraf}
\title{Round Robin Allele Frequencies}
\usage{
rraf(gid, res = "list", by_pop = FALSE, correction = TRUE)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{res}{Either "list" (default), "vector", or "data.frame".}

\item{by_pop}{When this is \code{TRUE}, the calculation will be done by
population. Defaults to \code{FALSE}}

\item{correction}{a logical indicating whether or not zero value allele
  frequencies should be set to 1/n (Default: \code{TRUE})}
}
\value{
a vector or list of allele frequencies
}
\description{
This function utilizes \code{\link{rrmlg}} to calculate multilocus genotypes
and then subsets each locus by the resulting MLGs to calculate the
round-robin allele frequencies used for pgen and psex.
}
\note{
When \code{by_pop = TRUE}, the output will be a matrix of allele
  frequencies.
}
\examples{
data(Pram)
# Round robin allele frequencies.
rraf(Pram)


\dontrun{

# Compare to without round robin:
PrLoc <- seploc(Pram, res = "mat") # get locus by matrix
lapply(PrLoc, colMeans, na.rm = TRUE)

# Without round robin, clone corrected:
Pcc <- clonecorrect(Pram, strata = NA) # indiscriminantly clone correct
PccLoc <- seploc(Pcc, res = "mat")
lapply(PccLoc, colMeans, na.rm = TRUE)

# Get vector output.
rraf(Pram, type = "vector")

# Get frequencies per population (matrix only)
rraf(Pram, by_pop = TRUE, correction = FALSE)

# Get data frame output and plot.
(Prdf <- rraf(Pram, res = "data.frame"))
library("ggplot2")
ggplot(Prdf, aes(y = allele, x = frequency)) +
  geom_point() +
  facet_grid(locus ~ ., scale = "free_y", space = "free")
}
}
\author{
Zhian N. Kamvar, Jonah C. Brooks, Stacey Hadfield
}
\seealso{
\code{\link{rrmlg}}, \code{\link{pgen}}, \code{\link{psex}}
}

