% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{snpclone-class}
\alias{snpclone}
\alias{snpclone-class}
\title{SNPclone class}
\description{
SNPclone is an S4 class that extends the \code{\linkS4class{genlight}}
from the \pkg{\link{adegenet}} package. It will have all of the same
attributes as the \code{\linkS4class{genlight}}, but it will contain two
extra slots that will help retain information about population hierarchies
and multilocus genotypes.
}
\details{
The snpclone class will allow for more optimized methods of clone
correcting and analyzing data over multiple levels of population hierarchy.
}
\section{Slots}{

\describe{
\item{\code{mlg}}{a vector representing multilocus genotypes for the data set.}

\item{\code{hierarchy}}{a data frame containing hierarchical levels.}
}}
\section{Extends}{

Class \code{"\linkS4class{genlight}"}, directly.
}
\author{
Zhian N. Kamvar
}
\seealso{
\code{\link{as.snpclone}} \code{\link{sethierarchy}} \code{\link{setpop}}
\code{\linkS4class{genlight}}
}

