% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaaMLGmethods.R
\docType{methods}
\name{unique,MLG-method}
\alias{duplicated,MLG-method}
\alias{unique,MLG-method}
\title{Unique and Duplicated implementations for MLG objects}
\usage{
\S4method{unique}{MLG}(x, incomparables = FALSE, ...)

\S4method{duplicated}{MLG}(x, incomparables = FALSE, ...)
}
\arguments{
\item{incomparables}{vector of values that cannot be compared}

\item{...}{options passed on to the base function \link{unique} or \link{duplicated}.}
}
\description{
internal use
}
\keyword{internal}

